requirements.txt
camel-ai
yfinance
camel-ai[tools]
tavily-pythona